functions moved to config.py
